removing user controller
