--- src/controllers/user.controller.js
@@ -1,294 +1,0 @@
-import sequelize from "#db/index.js";
-import User from "#models/user.model.js";
-import College from "#models/college.model.js";
-import { ApiError } from "#utils/api-handler/error.js";
-import { ApiResponse } from "#utils/api-handler/response.js";
-import { ApiHandler } from "#utils/api-handler/handler.js";
-import asyncLocalStorage from "#utils/context.js";
-
-import bcrypt from "bcrypt";
-import { Op } from "sequelize";
-
-import jwt from "jsonwebtoken";
-
-function generateAccessToken(user_id, email) {
-  const { session_id } = asyncLocalStorage.getStore();
-  return jwt.sign({ user_id, email, session_id }, process.env.TOKEN_SECRET, {
-    expiresIn: process.env.ACCESS_TOKEN_EXPIRY,
-  });
-}
-
-const onboardUsers = async (req, res) => {
-  const t = await sequelize.transaction();
-  try {
-    const { name, email, password, role, college_id } = req.body;
-
-    // Validate all fields
-    if (
-      [name, email, password, role, college_id].some(
-        (field) => !field || String(field).trim() === ""
-      )
-    ) {
-      return res.status(400).json({
-        status: "FAILURE",
-        responseMsg: "All_Feild are required",
-      });
-    }
-
-    // Check for existing email
-    const existingUser = await User.findOne({ where: { email } });
-    if (existingUser) {
-      return res.status(409).json({
-        status: "FAILURE",
-        responseMsg: "USER_ALREADY_EXISTS",
-      });
-    }
-
-    // Check for existing email
-    const existingCollege = await College.findOne({
-      where: { id: college_id },
-    });
-    if (!existingCollege) {
-      return res.status(409).json({
-        status: "FAILURE",
-        responseMsg: "COLLAGE_dOES_NOT_EXISTS",
-      });
-    }
-
-    if (req.user.role === "SUPERADMIN") {
-      // Hash password
-      const password_hash = await bcrypt.hash(password, 10);
-
-      // Create user with entity_id from college_id
-      const user = await User.create({
-        name,
-        email,
-        password_hash,
-        entity_id: college_id,
-        role,
-        active: true,
-      });
-
-      await t.commit();
-      return res.status(200).json({
-        status: "SUCCESS",
-        responseMsg: "USER_ONBOARDED",
-        payload: {
-          id: user.id,
-          name: user.name,
-          email: user.email,
-          role: user.role,
-          active: user.active,
-        },
-      });
-    } else {
-      await t.rollback();
-      return res.status(401).json({
-        status: "FAILURE",
-        responseMsg: "Unauthorized request",
-      });
-    }
-  } catch {
-    // console.error("Register error:", error);
-    await t.rollback();
-    return res.status(500).json({
-      status: "FAILURE",
-      responseMsg: "INTERNAL_SERVER_ERROR",
-    });
-  }
-};
-
-const loginUser = ApiHandler(async (req, res) => {
-  const { email, username, password } = req.body;
-
-  if (!email && !username) {
-    throw new ApiError(
-      400,
-      "AUTHENTICATION_FAILED",
-      "email or username is required"
-    );
-  }
-
-  const conditions = [];
-  if (email?.trim()) conditions.push({ email });
-  if (username?.trim()) conditions.push({ username });
-
-  const user = await User.findOne({
-    where: {
-      [Op.or]: conditions,
-    },
-  });
-
-  if (!user) {
-    throw new ApiError(404, "AUTHENTICATION_FAILED", "User not found");
-  }
-
-  const isMatch = await bcrypt.compare(password, user.password_hash);
-
-  if (!isMatch) {
-    throw new ApiError(400, "AUTHENTICATION_FAILED", "Invalid credentials");
-  }
-
-  const token = generateAccessToken(user.id, user.email);
-
-  return res.status(200).json(
-    new ApiResponse("LOGIN_SUCCESSFUL", {
-      token,
-      user: {
-        name: user.name,
-        email: user.email,
-        role: user.role,
-      },
-    })
-  );
-});
-
-const getAllAdmins = async (req, res) => {
-  try {
-    const page = parseInt(req.query.page) || 1;
-    const limit = parseInt(req.query.limit) || 10;
-    const offset = (page - 1) * limit;
-    const collegeId = req.query.college_id;
-    const role = req.query.role.toUpperCase();
-
-    // Authorization check
-    if (req.user.role !== "SUPERADMIN") {
-      return res.status(401).json({
-        status: "FAILURE",
-        responseMsg: "Unauthorized request",
-      });
-    }
-
-    // Fetch all users with role = 'ADMIN'
-    const { rows: admins, count: total } = await User.findAndCountAll({
-      where: { role, entity_id: collegeId },
-      offset,
-      limit,
-      attributes: ["id", "name", "email", "role", "active"],
-      order: [["created_at", "ASC"]],
-    });
-
-    return res.status(200).json({
-      status: "SUCCESS",
-      responseMsg: "ADMINS_FETCHED",
-      payload: {
-        total,
-        page,
-        limit,
-        totalPages: Math.ceil(total / limit),
-        users: admins,
-      },
-    });
-  } catch (error) {
-    console.error("Error fetching admins:", error);
-    return res.status(500).json({
-      status: "FAILURE",
-      responseMsg: "INTERNAL_SERVER_ERROR",
-    });
-  }
-};
-
-const studentRegistration = async (req, res) => {
-  const t = await sequelize.transaction();
-  try {
-    const { name, email, password } = req.body;
-
-    // Validate all fields
-    if (
-      [name, email, password].some(
-        (field) => !field || String(field).trim() === ""
-      )
-    ) {
-      return res.status(400).json({
-        status: "FAILURE",
-        responseMsg: "All_Feild are required",
-      });
-    }
-
-    // Check for existing email
-    const existingUser = await User.findOne({ where: { email } });
-    if (existingUser) {
-      return res.status(409).json({
-        status: "FAILURE",
-        responseMsg: "USER_ALREADY_EXISTS",
-      });
-    }
-
-    // Hashing Password
-    const password_hash = await bcrypt.hash(password, 10);
-    const user = await User.create({
-      name,
-      email,
-      password_hash,
-      role: "STUDENT",
-      active: true,
-    });
-
-    await t.commit();
-    return res.status(200).json({
-      status: "SUCCESS",
-      responseMsg: "STUDENT_REGISTERED",
-      payload: {
-        id: user.id,
-        role: user.role,
-      },
-    });
-  } catch {
-    await t.rollback();
-    return res.status(500).json({
-      status: "FAILURE",
-      responseMsg: "INTERNAL_SERVER_ERROR",
-    });
-  }
-};
-
-const getAllStudents = async (req, res) => {
-  try {
-    const page = parseInt(req.query.page) || 1;
-    const limit = parseInt(req.query.limit) || 10;
-    const offset = (page - 1) * limit;
-
-    // Authorization check
-    if (!(req.user.role === "ADMIN" || req.user.role === "SUPERADMIN")) {
-      return res.status(401).json({
-        status: "FAILURE",
-        responseMsg: "Unauthorized request",
-      });
-    }
-
-    // Fetch all users with role = 'ADMIN'
-    const { rows: students, count: total } = await User.findAndCountAll({
-      where: { role: "STUDENT" },
-      offset,
-      limit,
-      attributes: ["id", "name", "email", "role", "active"],
-      order: [["created_at", "ASC"]],
-    });
-
-    return res.status(200).json({
-      status: "SUCCESS",
-      responseMsg: "STUDENTS_FETCHED",
-      payload: {
-        total,
-        page,
-        limit,
-        totalPages: Math.ceil(total / limit),
-        users: students,
-      },
-    });
-  } catch (error) {
-    console.error("Error fetching admins:", error);
-    return res.status(500).json({
-      status: "FAILURE",
-      responseMsg: "INTERNAL_SERVER_ERROR",
-    });
-  }
-};
-
-export {
-  onboardUsers,
-  loginUser,
-  getAllAdmins,
-  studentRegistration,
-  getAllStudents,
-};
